requirements.txt
# Bybit API Client
pybit==5.6.2

# Async support
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
aiosqlite==0.19.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Data handling
pandas==2.1.3
numpy==1.26.2

# Logging
structlog==23.2.0
colorama==0.4.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Rate limiting
ratelimit==2.2.1

# HTTP requests
requests==2.31.0
httpx==0.25.1

# WebSocket
websockets==12.0

# CORS support
fastapi-cors==0.0.6
